requirements.txt
streamlit
langchain
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
nltk
numpy

langchain-openai

langchain-core
python-dotenv
chromadb

pypdf
numpy==1.26.4
nltk

tf-keras
torch
transformers
